backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
python-dotenv
pytest
pytest-asyncio
httpx
pytest-cov
